console-server: Add a timeout facility for pollers
This change adds a timeout callback for pollers, allowing processing
to occur outside of the usual file descriptor events.

We add a timeout callback argument to console_poller_register:

  struct poller *console_poller_register(struct console *console,
  		struct handler *handler,
  		poller_event_fn_t event_fn,
  		poller_timeout_fn_t timeout_fn,
  		int fd, int events, void *data)

Pollers can then schedule a future call via:

  void console_poller_set_timeout(struct console *console,
		struct poller *poller, struct timeval *tv)

Otherwise, the semantics for the timeout callback are the same as for
the event callback.

Currently untested; do not merge.

Change-Id: Id7caf3b6a2b8f5d6990b4bc0aab59763cc52cee1
Signed-off-by: Jeremy Kerr <jk@ozlabs.org>