socket handler: allow batching of sent data
Currently, if the UART device gives us small amounts of data at a time,
and a connected socket is not blocked on write(), then we'll end up
sending those small chunks of data through the socket. This may result
in large overheads in sending these small chunks in many syscalls.

This change uses the timeout failicty to batch writes.

Currently untested; do not merge.

Change-Id: Iae2d05f0021f3e68d3eb4990f2051e9abe949680
Signed-off-by: Jeremy Kerr <jk@ozlabs.org>